+++ galsim/gsobject.py
@@ -333,7 +333,7 @@ class GSObject(object):
         """Returns an estimate of the maximum surface brightness of the object.
 
         Some profiles will return the exact peak SB, typically equal to the value of
-        obj.xValue(obj.centroid()).  However, not all profiles (e.g. Convolution) know how
+        obj.xValue(obj.centroid()).  However, not all profiles (e.g. Convolution) know how to
         calculate this value without just drawing the image and checking what the maximum value is.
         Clearly, this would be inefficient, so in these cases, some kind of estimate is returned,
         which will generally be conservative on the high side.
@@ -1159,7 +1159,7 @@ class GSObject(object):
         Given the periodicity implicit in the use of FFTs, there can occasionally be artifacts due
         to wrapping at the edges, particularly for objects that are quite extended (e.g., due to
         the nature of the radial profile). See `help(galsim.GSParams)` for parameters that you can
-        use to reduce the level of these artificats, in particular `folding_threshold` may be
+        use to reduce the level of these artifacts, in particular `folding_threshold` may be
         helpful if you see such artifacts in your images.
 
         @param image        If provided, this will be the image on which to draw the profile.
@@ -1401,7 +1401,7 @@ class GSObject(object):
                             Note: The flux will be added to any flux already on the image,
                             so this corresponds to the add_to_image=True option in drawImage.
 
-        @returns The total flux drawin inside the image bounds.
+        @returns The total flux drawn inside the image bounds.
         """
         return self.SBProfile.draw(image.image, image.scale)
 
@@ -1409,7 +1409,7 @@ class GSObject(object):
         """Return a good size to use for drawing this profile.
 
         The size will be large enough to cover most of the flux of the object.  Specifically,
-        at least (1-gsparasm.folding_threshold) (i.e. 99.5% by default) of the flux should fall
+        at least (1-gsparams.folding_threshold) (i.e. 99.5% by default) of the flux should fall
         in the image.
 
         Also, the returned size is always an even number, which is usually desired in practice.
@@ -1446,7 +1446,7 @@ class GSObject(object):
                             Note: The flux will be added to any flux already on the image,
                             so this corresponds to the add_to_image=True option in drawImage.
 
-        @returns The total flux drawin inside the image bounds.
+        @returns The total flux drawn inside the image bounds.
         """
         if wmult != 1.: # pragma: no cover
             from .deprecated import depr
@@ -1470,10 +1470,10 @@ class GSObject(object):
         # Make sure we hit the minimum size specified in the gsparams.
         N = max(N, self.gsparams.minimum_fft_size)
 
-        dk = 2.*np.pi / N;
+        dk = 2.*np.pi / N
 
         if N*dk/2 > self.maxK():
-            Nk = N;
+            Nk = N
         else:
             # There will be aliasing.  Make a larger image and then wrap it.
             Nk = int(np.ceil(self.maxK()/dk)) * 2
@@ -1499,7 +1499,7 @@ class GSObject(object):
 
         # Add (a portion of) this to the original image.
         image.image += real_image.subImage(image.bounds)
-        added_photons = real_image.subImage(image.bounds).array.sum();
+        added_photons = real_image.subImage(image.bounds).array.sum()
 
         return added_photons
 
@@ -1594,18 +1594,18 @@ class GSObject(object):
             pd = galsim.PoissonDeviate(rng, mean)
             pd_val = pd() - mean + flux
             ratio = pd_val / flux
-            g *= ratio;
-            mod_flux *= ratio;
+            g *= ratio
+            mod_flux *= ratio
 
         if n_photons == 0.:
-            n_photons = mod_flux;
+            n_photons = mod_flux
             if max_extra_noise > 0.:
                 gfactor = 1. + max_extra_noise / self.SBProfile.maxSB()
-                n_photons /= gfactor;
-                g *= gfactor;
+                n_photons /= gfactor
+                g *= gfactor
 
         # Make n_photons an integer.
-        iN = int(n_photons + 0.5);
+        iN = int(n_photons + 0.5)
 
         if iN <= 0:  # pragma: no cover
             import warnings
@@ -1688,8 +1688,7 @@ class GSObject(object):
             else: poisson_flux = False
 
         # Check that either n_photons is set to something or flux is set to something
-        if (n_photons == 0. and self.getFlux() == 1.
-            and area == 1. and exptime == 1.): # pragma: no cover
+        if n_photons == 0. and self.getFlux() == 1.:
             import warnings
             warnings.warn(
                     "Warning: drawImage for object with flux == 1, area == 1, and "
@@ -1732,7 +1731,7 @@ class GSObject(object):
             thisN = min(maxN, Nleft)
 
             try:
-                phot_array = self.SBProfile.shoot(thisN, ud);
+                phot_array = self.SBProfile.shoot(thisN, ud)
             except RuntimeError:  # pragma: no cover
                 # Give some extra explanation as a warning, then raise the original exception
                 # so the traceback shows as much detail as possible.
@@ -1742,12 +1741,12 @@ class GSObject(object):
                     "Deconvolve or is a compound including one or more Deconvolve objects.")
                 raise
 
-            phot_array.scaleFlux(g * thisN / Ntot);
+            phot_array.scaleFlux(g * thisN / Ntot)
 
             added_flux += sensor.accumulate(phot_array, image)
             Nleft -= thisN;
 
-        return added_flux;
+        return added_flux
 
 
     def drawKImage(self, image=None, nx=None, ny=None, bounds=None, scale=None,
@@ -1768,7 +1767,7 @@ class GSObject(object):
 
         @param image        If provided, this will be the ImageC onto which to draw the k-space
                             image.  If `image` is None, then an automatically-sized image will be
-                            created.  If is is given, but its bounds are undefined, then it
+                            created.  If `image` is given, but its bounds are undefined, then it
                             will be resized appropriately based on the profile's size.
                             [default: None]
         @param nx           If provided and `image` is None, use to set the x-direction size of the